@@ -1,38 +1,51 @@
 module github.com/google/googet/v2
 
-go 1.19
+go 1.23.0
+
+toolchain go1.23.4
 
 require (
-	cloud.google.com/go/storage v1.15.0
-	github.com/StackExchange/wmi v0.0.0-20210224194228-fe8f1750fd46
+	cloud.google.com/go/storage v1.28.1
 	github.com/blang/semver v3.5.1+incompatible
-	github.com/dustin/go-humanize v1.0.0
-	github.com/go-yaml/yaml v2.1.0+incompatible
+	github.com/dustin/go-humanize v1.0.1
+	github.com/google/go-cmp v0.7.0
 	github.com/google/logger v1.1.1
 	github.com/google/subcommands v1.2.0
 	github.com/olekukonko/tablewriter v0.0.5
-	golang.org/x/oauth2 v0.0.0-20210427180440-81ed05c6b58c
-	golang.org/x/sys v0.5.0
-	google.golang.org/api v0.46.0
+	github.com/yusufpapurcu/wmi v1.2.4
+	golang.org/x/oauth2 v0.7.0
+	golang.org/x/sys v0.31.0
+	google.golang.org/api v0.114.0
+	gopkg.in/yaml.v3 v3.0.1
+	modernc.org/sqlite v1.37.0
 )
 
 require (
-	cloud.google.com/go v0.81.0 // indirect
-	github.com/go-ole/go-ole v1.2.5 // indirect
+	cloud.google.com/go v0.110.0 // indirect
+	cloud.google.com/go/compute v1.19.1 // indirect
+	cloud.google.com/go/compute/metadata v0.2.3 // indirect
+	cloud.google.com/go/iam v0.13.0 // indirect
+	github.com/go-ole/go-ole v1.2.6 // indirect
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/googleapis/gax-go/v2 v2.0.5 // indirect
-	github.com/jstemmer/go-junit-report v0.9.1 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/googleapis/enterprise-certificate-proxy v0.2.3 // indirect
+	github.com/googleapis/gax-go/v2 v2.7.1 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.12 // indirect
+	github.com/ncruces/go-strftime v0.1.9 // indirect
+	github.com/remyoudompheng/bigfft v0.0.0-20230129092748-24d4a6f8daec // indirect
 	github.com/rivo/uniseg v0.2.0 // indirect
-	go.opencensus.io v0.23.0 // indirect
-	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
-	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
-	golang.org/x/net v0.7.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
-	golang.org/x/tools v0.1.12 // indirect
+	go.opencensus.io v0.24.0 // indirect
+	golang.org/x/exp v0.0.0-20250305212735-054e65f0b394 // indirect
+	golang.org/x/net v0.37.0 // indirect
+	golang.org/x/text v0.23.0 // indirect
+	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/genproto v0.0.0-20210506142907-4a47615972c2 // indirect
-	google.golang.org/grpc v1.37.0 // indirect
-	google.golang.org/protobuf v1.26.0 // indirect
+	google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1 // indirect
+	google.golang.org/grpc v1.56.3 // indirect
+	google.golang.org/protobuf v1.33.0 // indirect
+	modernc.org/libc v1.62.1 // indirect
+	modernc.org/mathutil v1.7.1 // indirect
+	modernc.org/memory v1.9.1 // indirect
 )
