@@ -25,6 +25,7 @@ import (
 	"github.com/google/googet/v2/client"
 	"github.com/google/googet/v2/download"
 	"github.com/google/googet/v2/goolib"
+	"github.com/google/googet/v2/settings"
 	"github.com/google/logger"
 	"github.com/google/subcommands"
 )
@@ -58,7 +59,12 @@ func (cmd *downloadCmd) Execute(ctx context.Context, flags *flag.FlagSet, _ ...i
 		logger.Fatal("No repos defined, create a .repo file or pass using the -sources flag.")
 	}
 
-	rm := client.AvailableVersions(ctx, repos, filepath.Join(rootDir, cacheDir), cacheLife, proxyServer)
+	downloader, err := client.NewDownloader(settings.ProxyServer)
+	if err != nil {
+		logger.Fatal(err)
+	}
+
+	rm := downloader.AvailableVersions(ctx, repos, settings.CacheDir(), settings.CacheLife)
 	exitCode := subcommands.ExitSuccess
 
 	dir := cmd.downloadDir
@@ -72,7 +78,7 @@ func (cmd *downloadCmd) Execute(ctx context.Context, flags *flag.FlagSet, _ ...i
 	for _, arg := range flags.Args() {
 		pi := goolib.PkgNameSplit(arg)
 		if pi.Ver == "" {
-			if _, err := download.Latest(ctx, pi.Name, dir, rm, archs, proxyServer); err != nil {
+			if _, _, err := download.Latest(ctx, pi.Name, dir, rm, settings.Archs, downloader); err != nil {
 				logger.Errorf("error downloading %s, %v", pi.Name, err)
 				exitCode = subcommands.ExitFailure
 			}
@@ -97,7 +103,7 @@ func (cmd *downloadCmd) Execute(ctx context.Context, flags *flag.FlagSet, _ ...i
 			exitCode = subcommands.ExitFailure
 			continue
 		}
-		if _, err := download.FromRepo(ctx, rs, repo, dir, proxyServer); err != nil {
+		if _, _, err := download.FromRepo(ctx, rs, repo, dir, downloader); err != nil {
 			logger.Errorf("error downloading %s.%s %s, %v", pi.Name, pi.Arch, pi.Ver, err)
 			exitCode = subcommands.ExitFailure
 			continue
