@@ -20,6 +20,7 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/google/googet/v2/settings"
 	"github.com/google/logger"
 	"github.com/google/subcommands"
 )
@@ -35,7 +36,7 @@ func (*listReposCmd) Usage() string {
 func (cmd *listReposCmd) SetFlags(f *flag.FlagSet) {}
 
 func (cmd *listReposCmd) Execute(_ context.Context, _ *flag.FlagSet, _ ...interface{}) subcommands.ExitStatus {
-	rfs, err := repos(filepath.Join(rootDir, repoDir))
+	rfs, err := repos(settings.RepoDir())
 	if err != nil {
 		logger.Fatal(err)
 	}
