@@ -39,6 +39,9 @@ var toRemove []string
 // minInstalled reports whether the package is installed at the given version or greater.
 func minInstalled(pi goolib.PackageInfo, state client.GooGetState) (bool, error) {
 	for _, p := range state {
+		if p.PackageSpec == nil {
+			continue
+		}
 		if p.PackageSpec.Name == pi.Name && (pi.Arch == "" || p.PackageSpec.Arch == pi.Arch) {
 			c, err := goolib.Compare(pi.Ver, p.PackageSpec.Version)
 			if err != nil {
@@ -67,7 +70,7 @@ func resolveConflicts(ps *goolib.PkgSpec, state *client.GooGetState) error {
 	return nil
 }
 
-func resolveReplacements(ctx context.Context, ps *goolib.PkgSpec, state *client.GooGetState, dbOnly bool, proxyServer string) error {
+func resolveReplacements(ctx context.Context, ps *goolib.PkgSpec, state *client.GooGetState, dbOnly bool, downloader *client.Downloader) error {
 	// Check for and remove any package this replaces.
 	// TODO(ajackura): Make sure no replacements are listed as
 	// dependencies or subdependancies.
@@ -82,14 +85,14 @@ func resolveReplacements(ctx context.Context, ps *goolib.PkgSpec, state *client.
 		}
 		deps, _ := remove.EnumerateDeps(pi, *state)
 		logger.Infof("%s replaces %s, removing", ps, pi)
-		if err := remove.All(ctx, pi, deps, state, dbOnly, proxyServer); err != nil {
+		if err := remove.All(ctx, pi, deps, state, dbOnly, downloader); err != nil {
 			return err
 		}
 	}
 	return nil
 }
 
-func installDeps(ctx context.Context, ps *goolib.PkgSpec, cache string, rm client.RepoMap, archs []string, state *client.GooGetState, dbOnly bool, proxyServer string) error {
+func installDeps(ctx context.Context, ps *goolib.PkgSpec, cache string, rm client.RepoMap, archs []string, state *client.GooGetState, dbOnly bool, downloader *client.Downloader) error {
 	logger.Infof("Resolving conflicts and dependencies for %s %s version %s", ps.Arch, ps.Name, ps.Version)
 	if err := resolveConflicts(ps, state); err != nil {
 		return err
@@ -116,39 +119,31 @@ func installDeps(ctx context.Context, ps *goolib.PkgSpec, cache string, rm clien
 		}
 		if c > -1 {
 			logger.Infof("Dependency found: %s.%s %s is available", pi.Name, arch, v)
-			if err := FromRepo(ctx, goolib.PackageInfo{Name: pi.Name, Arch: arch, Ver: v}, repo, cache, rm, archs, state, dbOnly, proxyServer); err != nil {
+			if err := FromRepo(ctx, goolib.PackageInfo{Name: pi.Name, Arch: arch, Ver: v}, repo, cache, rm, archs, state, dbOnly, downloader); err != nil {
 				return err
 			}
 			ins = true
 		}
 		if !ins {
-			return fmt.Errorf("cannot resolve dependancy, %s.%s version %s or greater not installed and not available in any repo", pi.Name, arch, ver)
+			return fmt.Errorf("cannot resolve dependency, %s.%s version %s or greater not installed and not available in any repo", pi.Name, arch, ver)
 		}
 	}
-	return resolveReplacements(ctx, ps, state, dbOnly, proxyServer)
+	return resolveReplacements(ctx, ps, state, dbOnly, downloader)
 }
 
 // FromRepo installs a package and all dependencies from a repository.
-func FromRepo(ctx context.Context, pi goolib.PackageInfo, repo, cache string, rm client.RepoMap, archs []string, state *client.GooGetState, dbOnly bool, proxyServer string) error {
-	ni, err := NeedsInstallation(pi, *state)
-	if err != nil {
-		return err
-	}
-	if !ni {
-		return nil
-	}
-
+func FromRepo(ctx context.Context, pi goolib.PackageInfo, repo, cache string, rm client.RepoMap, archs []string, state *client.GooGetState, dbOnly bool, downloader *client.Downloader) error {
 	logger.Infof("Starting install of %s.%s.%s", pi.Name, pi.Arch, pi.Ver)
 	fmt.Printf("Installing %s.%s.%s and dependencies...\n", pi.Name, pi.Arch, pi.Ver)
 	rs, err := client.FindRepoSpec(pi, rm[repo])
 	if err != nil {
 		return err
 	}
-	if err := installDeps(ctx, rs.PackageSpec, cache, rm, archs, state, dbOnly, proxyServer); err != nil {
+	if err := installDeps(ctx, rs.PackageSpec, cache, rm, archs, state, dbOnly, downloader); err != nil {
 		return err
 	}
 
-	dst, err := download.FromRepo(ctx, rs, repo, cache, proxyServer)
+	dst, pkgURL, err := download.FromRepo(ctx, rs, repo, cache, downloader)
 	if err != nil {
 		return err
 	}
@@ -166,7 +161,7 @@ func FromRepo(ctx context.Context, pi goolib.PackageInfo, repo, cache string, rm
 
 	state.Add(client.PackageState{
 		SourceRepo:     repo,
-		DownloadURL:    strings.TrimSuffix(repo, filepath.Base(repo)) + rs.Source,
+		DownloadURL:    pkgURL,
 		Checksum:       rs.Checksum,
 		LocalPath:      dst,
 		PackageSpec:    rs.PackageSpec,
@@ -176,24 +171,22 @@ func FromRepo(ctx context.Context, pi goolib.PackageInfo, repo, cache string, rm
 }
 
 // FromDisk installs a local .goo file.
-func FromDisk(arg, cache string, state *client.GooGetState, dbOnly, ri bool) error {
+func FromDisk(arg, cache string, state *client.GooGetState, dbOnly, ri bool) (client.GooGetState, error) {
 	if _, err := oswrap.Stat(arg); err != nil {
-		return err
+		return nil, err
 	}
-
 	zs, err := extractSpec(arg)
 	if err != nil {
-		return fmt.Errorf("error extracting spec file: %v", err)
+		return nil, fmt.Errorf("error extracting spec file: %v", err)
 	}
-
 	if !ri {
 		ni, err := NeedsInstallation(goolib.PackageInfo{Name: zs.Name, Arch: zs.Arch, Ver: zs.Version}, *state)
 		if err != nil {
-			return err
+			return nil, err
 		}
 		if !ni {
 			fmt.Printf("%s.%s.%s or a newer version is already installed on the system\n", zs.Name, zs.Arch, zs.Version)
-			return nil
+			return nil, nil
 		}
 	}
 
@@ -201,67 +194,62 @@ func FromDisk(arg, cache string, state *client.GooGetState, dbOnly, ri bool) err
 	fmt.Printf("Installing %s %s...\n", zs.Name, zs.Version)
 
 	if err := resolveConflicts(zs, state); err != nil {
-		return err
+		return nil, err
 	}
 	for p, ver := range zs.PkgDependencies {
 		pi := goolib.PkgNameSplit(p)
 		mi, err := minInstalled(goolib.PackageInfo{Name: pi.Name, Arch: pi.Arch, Ver: ver}, *state)
 		if err != nil {
-			return err
+			return nil, err
 		}
 		if mi {
 			logger.Infof("Dependency met: %s.%s with version greater than %s installed", pi.Name, pi.Arch, ver)
 			continue
 		}
-		return fmt.Errorf("package dependency %s %s (min version %s) not installed", pi.Name, pi.Arch, ver)
+		return nil, fmt.Errorf("package dependency %s %s (min version %s) not installed", pi.Name, pi.Arch, ver)
 	}
 	for _, pkg := range zs.Replaces {
 		pi := goolib.PkgNameSplit(pkg)
 		ins, err := minInstalled(goolib.PackageInfo{Name: pi.Name, Arch: pi.Arch, Ver: pi.Ver}, *state)
 		if err != nil {
-			return err
+			return nil, err
 		}
 		if ins {
-			return fmt.Errorf("cannot install, replaces installed package, remove first then try installation again: %s", pi)
+			return nil, fmt.Errorf("cannot install, replaces installed package, remove first then try installation again: %s", pi)
 		}
 	}
 
 	dst := filepath.Join(cache, goolib.PackageInfo{Name: zs.Name, Arch: zs.Arch, Ver: zs.Version}.PkgName())
 	if err := copyPkg(arg, dst); err != nil {
-		return err
+		return nil, err
 	}
 
 	insFiles, err := installPkg(dst, zs, dbOnly)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	if ri {
 		logger.Infof("Reinstallation of %q, version %q completed", zs.Name, zs.Version)
 		fmt.Printf("Reinstallation of %s completed\n", zs.Name)
-		return nil
+		return nil, nil
 	}
 
 	logger.Infof("Installation of %q, version %q completed", zs.Name, zs.Version)
 	fmt.Printf("Installation of %s completed\n", zs.Name)
 
-	// Clean up old version, if applicable.
-	pi := goolib.PackageInfo{Name: zs.Name, Arch: zs.Arch, Ver: ""}
-	cleanOld(state, pi, insFiles, dbOnly)
-
-	state.Add(client.PackageState{
+	return client.GooGetState{client.PackageState{
 		LocalPath:      dst,
 		PackageSpec:    zs,
 		InstalledFiles: insFiles,
-	})
-	return nil
+	}}, nil
 }
 
 // Reinstall reinstalls and optionally redownloads, a package.
-func Reinstall(ctx context.Context, ps client.PackageState, state client.GooGetState, rd bool, proxyServer string) error {
-	pi := goolib.PackageInfo{Name: ps.PackageSpec.Name, Arch: ps.PackageSpec.Arch, Ver: ps.PackageSpec.Version}
-	logger.Infof("Starting reinstall of %s.%s, version %s", pi.Name, pi.Arch, pi.Ver)
-	fmt.Printf("Reinstalling %s.%s %s and dependencies...\n", pi.Name, pi.Arch, pi.Ver)
+func Reinstall(ctx context.Context, ps client.PackageState, rd bool, downloader *client.Downloader) error {
+	spec := ps.PackageSpec
+	logger.Infof("Starting reinstall of %s.%s, version %s", spec.Name, spec.Arch, spec.Version)
+	fmt.Printf("Reinstalling %s.%s %s and dependencies...\n", spec.Name, spec.Arch, spec.Version)
 
 	// Fix for package install by older versions of GooGet.
 	if ps.LocalPath == "" && ps.UnpackDir != "" {
@@ -269,7 +257,7 @@ func Reinstall(ctx context.Context, ps client.PackageState, state client.GooGetS
 	}
 
 	if ps.LocalPath == "" {
-		return fmt.Errorf("local path not referenced in state file for %s.%s.%s. Cannot redownload", pi.Name, pi.Arch, pi.Ver)
+		return fmt.Errorf("local path not referenced in state file for %s.%s.%s. Cannot redownload", spec.Name, spec.Arch, spec.Version)
 	}
 
 	f, err := os.Open(ps.LocalPath)
@@ -277,7 +265,7 @@ func Reinstall(ctx context.Context, ps client.PackageState, state client.GooGetS
 		return err
 	}
 	if os.IsNotExist(err) {
-		logger.Infof("Local package does not exist for %s.%s.%s, redownloading...", pi.Name, pi.Arch, pi.Ver)
+		logger.Infof("Local package does not exist for %s.%s.%s, redownloading...", spec.Name, spec.Arch, spec.Version)
 		rd = true
 	}
 	// Force redownload if checksum does not match.
@@ -290,9 +278,9 @@ func Reinstall(ctx context.Context, ps client.PackageState, state client.GooGetS
 
 	if rd {
 		if ps.DownloadURL == "" {
-			return fmt.Errorf("can not redownload %s.%s.%s, DownloadURL not saved", pi.Name, pi.Arch, pi.Ver)
+			return fmt.Errorf("can not redownload %s.%s.%s, DownloadURL not saved", spec.Name, spec.Arch, spec.Version)
 		}
-		if err := download.Package(ctx, ps.DownloadURL, ps.LocalPath, ps.Checksum, proxyServer); err != nil {
+		if err := download.Package(ctx, ps.DownloadURL, ps.LocalPath, ps.Checksum, downloader); err != nil {
 			return fmt.Errorf("error redownloading package: %v", err)
 		}
 	}
@@ -301,8 +289,8 @@ func Reinstall(ctx context.Context, ps client.PackageState, state client.GooGetS
 		return fmt.Errorf("error reinstalling package: %v", err)
 	}
 
-	logger.Infof("Reinstallation of %s.%s, version %s completed", pi.Name, pi.Arch, pi.Ver)
-	fmt.Printf("Reinstallation of %s.%s %s completed\n", pi.Name, pi.Arch, pi.Ver)
+	logger.Infof("Reinstallation of %s.%s, version %s completed", spec.Name, spec.Arch, spec.Version)
+	fmt.Printf("Reinstallation of %s.%s %s completed\n", spec.Name, spec.Arch, spec.Version)
 	return nil
 }
 
@@ -332,10 +320,10 @@ func copyPkg(src, dst string) (retErr error) {
 // NeedsInstallation checks if a package version needs installation.
 func NeedsInstallation(pi goolib.PackageInfo, state client.GooGetState) (bool, error) {
 	for _, p := range state {
-		if p.PackageSpec.Name == pi.Name {
-			if p.PackageSpec.Arch != pi.Arch {
-				continue
-			}
+		if p.PackageSpec == nil {
+			continue
+		}
+		if p.PackageSpec.Name == pi.Name && p.PackageSpec.Arch == pi.Arch {
 			c, err := goolib.Compare(p.PackageSpec.Version, pi.Ver)
 			if err != nil {
 				return true, err
