@@ -17,22 +17,70 @@ limitations under the License.
 package system
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
 	"runtime"
+	"strings"
+	"time"
 
-	"github.com/StackExchange/wmi"
+	"github.com/google/googet/v2/client"
 	"github.com/google/googet/v2/goolib"
 	"github.com/google/googet/v2/oswrap"
 	"github.com/google/logger"
+	"github.com/yusufpapurcu/wmi"
 	"golang.org/x/sys/windows/registry"
 )
 
 const uninstallBase = `SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\`
 
-var msiSuccessCodes = []int{1641, 3010}
+// Regex taken from Winget uninstaller
+// https://github.com/microsoft/winget-cli/blob/6ea13623e5e4b870b81efeea9142d15a98dd4208/src/AppInstallerCommonCore/NameNormalization.cpp#L262
+var (
+	programNameReg = []string{
+		"PrefixParens",
+		"EmptyParens",
+		"Version",
+		"TrailingSymbols",
+		"LeadingSymbols",
+		"FilePathParens",
+		"FilePathQuotes",
+		"FilePath",
+		"VersionLetter",
+		"VersionLetterDelimited",
+		"En",
+		"NonNestedBracket",
+		"BracketEnclosed",
+		"URIProtocol",
+	}
+	publisherNameReg = []string{
+		"VersionDelimited",
+		"Version",
+		"NonNestedBracket",
+		"BracketEnclosed",
+		"URIProtocol",
+	}
+	regex = map[string]string{
+		"PrefixParens":           `(^\(.*?\))`,
+		"EmptyParens":            `((\(\s*\)|\[\s*\]|"\s*"))`,
+		"Version":                `(?:^)|(?:P|V|R|VER|VERSI(?:O|Ó)N|VERSÃO|VERSIE|WERSJA|BUILD|RELEASE|RC|SP)(?:\P{L}|\P{L}\p{L})?(\p{Nd}|\.\p{Nd})+(?:RC|B|A|R|V|SP)?\p{Nd}?`,
+		"TrailingSymbols":        `([^\p{L}\p{Nd}]+$)`,
+		"LeadingSymbols":         `(^[^\p{L}\p{Nd}]+)`,
+		"FilePathParens":         `(\([CDEF]:\\(.+?\\)*[^\s]*\\?\))`,
+		"FilePathQuotes":         `("[CDEF]:\\(.+?\\)*[^\s]*\\?")`,
+		"FilePath":               `(((INSTALLED\sAT|IN)\s)?[CDEF]:\\(.+?\\)*[^\s]*\\?)`,
+		"VersionLetter":          `((?:^\p{L})(?:(?:V|VER|VERSI(?:O|Ó)N|VERSÃO|VERSIE|WERSJA|BUILD|RELEASE|RC|SP)\P{L})?\p{Lu}\p{Nd}+(?:[\p{Po}\p{Pd}\p{Pc}]\p{Nd}+)+)`,
+		"VersionLetterDelimited": `((?:^\p{L})(?:(?:V|VER|VERSI(?:O|Ó)N|VERSÃO|VERSIE|WERSJA|BUILD|RELEASE|RC|SP)\P{L})?\p{Lu}\p{Nd}+(?:[\p{Po}\p{Pd}\p{Pc}]\p{Nd}+)+)`,
+		"En":                     `(\sEN\s*$)`,
+		"NonNestedBracket":       `(\([^\(\)]*\)|\[[^\[\]]*\])`,
+		"BracketEnclosed":        `((?:\p{Ps}.*\p{Pe}|".*"))`,
+		"URIProtocol":            `((?:^\p{L})(?:http[s]?|ftp):\/\/)`,
+	}
+	msiSuccessCodes = []int{1641, 3010}
+)
 
 func addUninstallEntry(dir string, ps *goolib.PkgSpec) error {
 	reg := uninstallBase + "GooGet - " + ps.Name
@@ -52,6 +100,7 @@ func addUninstallEntry(dir string, ps *goolib.PkgSpec) error {
 		{"InstallLocation", dir},
 		{"DisplayVersion", ps.Version},
 		{"DisplayName", "GooGet - " + ps.Name},
+		{"InstallDate", time.Now().Format("20060102")},
 	}
 	for _, re := range table {
 		if err := k.SetStringValue(re.name, re.value); err != nil {
@@ -65,6 +114,138 @@ func removeUninstallEntry(name string) error {
 	reg := uninstallBase + "GooGet - " + name
 	logger.Infof("Removing uninstall entry %q from registry.", reg)
 	return registry.DeleteKey(registry.LOCAL_MACHINE, reg)
+}
+
+// AppAssociation locates and returns registry entry and name of installed application.
+func AppAssociation(ps *goolib.PkgSpec, installSource string) (string, string) {
+	// Packages with files are portable and don't have actual installers.
+	if ps.Files != nil {
+		return "", ""
+	}
+	var productroots = []string{
+		`SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall\`,
+		`SOFTWARE\WOW6432Node\Microsoft\Windows\CurrentVersion\Uninstall\`,
+	}
+	for _, productroot := range productroots {
+		k, err := registry.OpenKey(registry.LOCAL_MACHINE, productroot, registry.ENUMERATE_SUB_KEYS)
+		if err != nil {
+			return "", ""
+		}
+		defer k.Close()
+		reg, err := k.ReadSubKeyNames(-1)
+		if err != nil {
+			return "", ""
+		}
+		for _, v := range reg {
+			productReg := fmt.Sprintf("%s%s", productroot, v)
+			q, err := registry.OpenKey(registry.LOCAL_MACHINE, productReg, registry.ALL_ACCESS)
+			if err != nil {
+				continue
+			}
+			defer q.Close()
+			displayName, _, err := q.GetStringValue("DisplayName")
+			if err != nil {
+				continue
+			}
+
+			if filepath.Ext(ps.Install.Path) == ".msi" && installSource != "" {
+				a, _, err := q.GetStringValue("InstallSource")
+				if err != nil {
+					// InstallSource not found, move on to next entry
+					continue
+				}
+				iS := strings.Split(installSource, "@")
+				if strings.Contains(a, iS[0]) {
+					name, _, err := q.GetStringValue("DisplayName")
+					if err != nil {
+						// UninstallString not found, move on to next entry
+						continue
+					}
+					return name, productReg
+				}
+			}
+			if displayName == ps.ExternalProgramName && ps.ExternalProgramName != "" {
+				return displayName, productReg
+			}
+			// TODO: Look into precompiling regex
+			var publisher, programName string
+			for _, v := range publisherNameReg {
+				re := regexp.MustCompile("(?i)" + regex[v])
+				publisher = re.ReplaceAllString(ps.Authors, "")
+			}
+			for _, v := range programNameReg {
+				re := regexp.MustCompile("(?i)" + regex[v])
+				programName = re.ReplaceAllString(ps.Name, "")
+			}
+			// Ignore empty and googet labeled pacakges
+			if displayName == "" || strings.Contains(displayName, "GooGet -") {
+				continue
+			}
+			// Check if Package name is in display name removing spaces
+			regPublisher, _, err := q.GetStringValue("Publisher")
+			if err != nil {
+				continue
+			}
+			if strings.Contains(strings.ToLower(strings.ReplaceAll(displayName, " ", "")), strings.ToLower(programName)) {
+				// Do an extra check for publisher for smaller package names and gain more confidence that we are returning the right package.
+				if len(programName) < 4 && !strings.Contains(strings.ToLower(strings.ReplaceAll(regPublisher, " ", "")), strings.ToLower(publisher)) {
+					return "", ""
+				}
+				return displayName, productReg
+			}
+			// Check if Package name is in display name removing dashes
+			if strings.Contains(strings.ToLower(strings.ReplaceAll(displayName, "-", "")), strings.ToLower(programName)) {
+				// Do an extra check for publisher for smaller package names and gain more confidence that we are returning the right package.
+				if len(programName) < 4 && !strings.Contains(strings.ToLower(strings.ReplaceAll(regPublisher, " ", "")), strings.ToLower(publisher)) {
+					return "", ""
+				}
+				return displayName, productReg
+			}
+			a, _, err := q.GetStringValue("InstallSource")
+			if err != nil {
+				// InstallSource not found, move on to next entry
+				continue
+			}
+			if installSource == "" {
+				return "", ""
+			}
+			iS := strings.Split(installSource, "@")
+			if strings.Contains(a, iS[0]) && installSource != "" {
+				return displayName, productReg
+			}
+		}
+	}
+	return "", ""
+}
+
+func uninstallString(regkey, extension string) string {
+	q, err := registry.OpenKey(registry.LOCAL_MACHINE, regkey, registry.ALL_ACCESS)
+	if err != nil {
+		logger.Error(err)
+		return ""
+	}
+	defer q.Close()
+	switch extension {
+	case "msi":
+		un, _, err := q.GetStringValue("UninstallString")
+		if err != nil {
+			logger.Error(err)
+			// UninstallString not found, move on
+			return ""
+		}
+		return un
+	default:
+		un, _, err := q.GetStringValue("QuietUninstallString")
+		if err != nil {
+			un, _, err = q.GetStringValue("UninstallString")
+			if err != nil {
+				// UninstallString not found, move on to next entry
+				return ""
+			}
+		}
+		return un
+	}
+
 }
 
 // Install performs a system specfic install given a package extraction directory and a PkgSpec struct.
@@ -99,6 +280,11 @@ func Install(dir string, ps *goolib.PkgSpec) error {
 		err = goolib.Run(exec.Command("wusa", args...), ec, out)
 	case ".exe":
 		err = goolib.Run(exec.Command(s, in.Args...), ec, out)
+	case ".msix", ".msixbundle":
+		// Add-AppxProvisionedPackage will install for all users.
+		installCmd := fmt.Sprintf("Add-AppxProvisionedPackage -online -PackagePath %v -SkipLicense", s)
+		args := append([]string{installCmd}, in.Args...)
+		err = goolib.Run(exec.Command("powershell", args...), ec, out)
 	default:
 		err = goolib.Exec(s, in.Args, in.ExitCodes, out)
 	}
@@ -114,15 +300,50 @@ func Install(dir string, ps *goolib.PkgSpec) error {
 }
 
 // Uninstall performs a system specfic uninstall given a packages PackageState.
-func Uninstall(dir string, ps *goolib.PkgSpec) error {
+func Uninstall(dir string, state *client.PackageState) error {
+	var filePath string
+	ps := state.PackageSpec
 	un := ps.Uninstall
+	r := regexp.MustCompile(`[^\s"]+|"([^"]*)"`)
+	// Automatically determine uninstall script if none is specified in spec.
 	if un.Path == "" {
-		return nil
+		switch filepath.Ext(ps.Install.Path) {
+		case ".msi":
+			u := uninstallString(state.InstalledApp.Reg, "msi")
+			if u == "" {
+				return nil
+			}
+			u = strings.ReplaceAll(u, `/I`, `/X`)
+			commands := r.FindAllString(u, -1)
+			un.Path = strings.Replace(commands[0], "\"", "", -1)
+			un.Args = []string{"/qn", "/norestart"}
+			un.Args = append(commands[1:], un.Args...)
+			filePath = un.Path
+		case ".msix", ".msixbundle":
+			un.Path = ps.Install.Path
+			filePath = un.Path
+		default:
+			u := uninstallString(state.InstalledApp.Reg, "")
+			commands := r.FindAllString(u, -1)
+			if len(commands) > 0 {
+				// Remove the quotes from the install string since we handle that below
+				un.Path = strings.Replace(commands[0], "\"", "", -1)
+				un.Args = commands[1:]
+				filePath = un.Path
+			}
+		}
+		if un.Path == "" {
+			return nil
+		}
 	}
-
 	logger.Infof("Running uninstall command: %q", un.Path)
 	// logging is only useful for failed uninstall
-	out, err := oswrap.Create(filepath.Join(dir, un.Path+".log"))
+	// Only append the directory if the folder structure doesn't exist
+	logPath := fmt.Sprintf("%s.log", un.Path)
+	if _, err := os.Stat(un.Path); errors.Is(err, os.ErrNotExist) {
+		logPath = filepath.Join(dir, logPath)
+	}
+	out, err := oswrap.Create(logPath)
 	if err != nil {
 		return err
 	}
@@ -131,18 +352,25 @@ func Uninstall(dir string, ps *goolib.PkgSpec) error {
 			logger.Error(err)
 		}
 	}()
-	s := filepath.Join(dir, un.Path)
+	if filePath == "" {
+		filePath = filepath.Join(dir, un.Path)
+	}
 	ec := append(msiSuccessCodes, un.ExitCodes...)
-	switch filepath.Ext(s) {
+	switch filepath.Ext(filePath) {
 	case ".msi":
 		msiLog := filepath.Join(dir, "msi_uninstall.log")
-		args := append([]string{"/x", s, "/qn", "/norestart", "/log", msiLog}, un.Args...)
+		args := append([]string{"/x", filePath, "/qn", "/norestart", "/log", msiLog}, un.Args...)
 		err = goolib.Run(exec.Command("msiexec", args...), ec, out)
 	case ".msu":
-		args := append([]string{s, "/uninstall", "/quiet", "/norestart"}, un.Args...)
+		args := append([]string{filePath, "/uninstall", "/quiet", "/norestart"}, un.Args...)
 		err = goolib.Run(exec.Command("wusa", args...), ec, out)
 	case ".exe":
-		err = goolib.Run(exec.Command(s, un.Args...), ec, out)
+		err = goolib.Run(exec.Command(filePath, un.Args...), ec, out)
+	case ".msix", ".msixbundle":
+		s := strings.Split(filepath.Base(filePath), "_")[0]
+		removeCmd := fmt.Sprintf(`Get-AppxProvisionedPackage -online | Where {$_.DisplayName -match "%v*"} | Remove-AppProvisionedPackage -online -AllUsers`, s)
+		args := append([]string{removeCmd}, un.Args...)
+		err = goolib.Run(exec.Command("powershell", args...), ec, out)
 	default:
 		err = goolib.Exec(filepath.Join(dir, un.Path), un.Args, un.ExitCodes, out)
 	}
@@ -188,6 +416,8 @@ func InstallableArchs() ([]string, error) {
 		return []string{"noarch", "x86_32", "x86_64"}, nil
 	case runtime.GOARCH == "arm":
 		return []string{"noarch", "arm"}, nil
+	case runtime.GOARCH == "arm64":
+		return []string{"noarch", "x86_32", "x86_64", "arm64"}, nil
 	default:
 		return nil, fmt.Errorf("runtime %s not supported", runtime.GOARCH)
 	}
